requirements.txt
spotipy>=2.23.0
ytmusicapi>=1.3.0
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
requests>=2.31.0
transformers>=4.36.0
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
yt-dlp>=2024.1.0
flask>=3.0.0
